Create first iteration of crystal-maze-prototype
My colleagues are planning on making a electronic game for a
children-in-need event at work. It will probably involve 10 large
buttons each with an LED. This project is to prototype the game that
someone suggested.

The prototype will be a webpage. I'm doing it in Rust and WASM following
the tutorial. The point of the prototype is to test if the suggested
game is fun and understandable.