Add the game code
It is being done in c, because it needs to run on an Arduino.

It is completely untested/unverified at this stage. My plan is to make
it into a WASM game again, and into an Arduino program.

I will be following this guide for the WASM:
https://developers.google.com/web/updates/2018/03/emscripting-a-c-library

I expect I will use [tinkercad](https://www.tinkercad.com) for the
arduino parts. I will probably screenshot the circuit, and commit it
that way.