Add a JavaScript version of the game
This will allow faster development without having to have an Arduino
handy.
The c code that the Arduino will use is compiled to WASM, which is run
by the JavaScript.

Next steps:
Reset should shuffle the button functions
Lay out the Arduino circuit in an online tool and code the integration
parts of the program.